refactor: - renamed files to follow styleguide
