+++ client/src/app.js
@@ -13,7 +13,7 @@ import '../../node_modules/toastr/build/toastr.min.js';
 
 import store from './store';
 import NavBarMain from './components/NavBarMain.jsx';
-import Footer from './components/footer.jsx';
+import Footer from './components/Footer.jsx';
 
 render(
   <Provider store={store}>
