+++ client/src/routes.js
@@ -5,15 +5,15 @@ import { createBrowserHistory } from 'history';
 import Signin from './components/Signin.jsx';
 import Signup from './components/Signup.jsx';
 import authRoute from './components/auth/authenticate.jsx';
-import Getevents from './components/Getevents.jsx';
-import Addevents from './components/Addevents.jsx';
+import GetEvents from './components/GetEvents.jsx';
+import AddEvent from './components/Addevent.jsx';
 import Center from './components/Center.jsx';
 import Landing from './components/Landing.jsx';
-import eventDetails from './components/eventDetails.jsx';
+import DetailEvent from './components/DetailEvent.jsx';
 import AddCenter from './components/AddCenter.jsx';
-import CenterDetails from './components/centerDetails.jsx';
-import EditCenter from './components/editCenter.jsx';
-import EditEvent from './components/editEvent.jsx';
+import CenterDetails from './components/CenterDetails.jsx';
+import EditCenter from './components/EditCenter.jsx';
+import EditEvent from './components/EditEvent.jsx';
 import GetMyEvents from './components/GetMyEvents.jsx';
 
 export const history = createBrowserHistory();
@@ -25,12 +25,12 @@ export const Routes = () => {
       <Route exact path="/getcenters" component={Center} />
       <Route exact path="/signin" component={Signin} />
       <Route exact path="/myevents" component={authRoute(GetMyEvents)} />
-      <Route exact path="/events" component={authRoute(Getevents)} />
-      <Route exact path="/addevents" component={authRoute(Addevents)} />
+      <Route exact path="/events" component={authRoute(GetEvents)} />
+      <Route exact path="/addevents" component={authRoute(AddEvent)} />
       <Route
         exact
         path="/eventdetails/:id"
-        component={authRoute(eventDetails)}
+        component={authRoute(DetailEvent)}
       />
       <Route
         exact
